# docs/api-spec.yml
openapi: 3.0.3
info:
  title: CinemaCore API
  description: >
    API Gateway specification for the CinemaCore Microservices ecosystem. 
    This acts as the single source of truth for frontend-backend contracts.
  version: 1.0.0
servers:
  - url: /api/v1
    description: API Gateway Entry Point

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    MovieCarteleraResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        titulo:
          type: string
        poster_path:
          type: string
        backdrop_path:
          type: string
        estado:
          type: string
        restriccion_edad:
          type: string
        fecha_estreno:
          type: string
          format: date
        trailers:
          type: array
          items:
            type: string
    MovieDetailResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        titulo:
          type: string
        sinopsis:
          type: string
        duracion_min:
          type: integer
        fecha_estreno:
          type: string
          format: date
        estado:
          type: string
        restriccion_edad:
          type: string
        poster_path:
          type: string
        backdrop_path:
          type: string
        director:
          type: string
        actores:
          type: array
          items:
            type: string
        generos:
          type: array
          items:
            type: string
        trailers:
          type: array
          items:
            type: string
paths:
  # ==========================================
  # AUTHENTICATION SERVICE
  # ==========================================
  /auth/register:
    post:
      summary: Registra un nuevo usuario cliente
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, contrasena, nombres, apellidos]
              properties:
                email:
                  type: string
                contrasena:
                  type: string
                nombres:
                  type: string
                apellidos:
                  type: string
      responses:
        '201':
          description: Usuario registrado exitosamente y auto-logueado
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  email:
                    type: string
                  roles:
                    type: array
                    items:
                      type: string
        '400':
          description: Errores de validación o correo duplicado (BadRequestException)

  /auth/login:
    post:
      summary: Inicia sesión y obtiene un JWT
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, contrasena]
              properties:
                email:
                  type: string
                contrasena:
                  type: string
      responses:
        '200':
          description: Login exitoso
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  email:
                    type: string
                  roles:
                    type: array
                    items:
                      type: string
        '401':
          description: Credenciales inválidas

  # ==========================================
  # MOVIE SERVICE (Curación de Catálogo)
  # ==========================================
  /movies/tmdb/search:
    get:
      summary: (Admin) Buscar en TMDB
      operationId: searchTmdbMovies
      tags: [Movies]
      security:
        - bearerAuth: []
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Lista de resultados básicos de TMDB
          
  /movies/tmdb/import/{tmdbId}:
    post:
      summary: (Admin) Importar al Cine Local
      description: Si la película ya existe (tmdbId), se actualizará su información externa manteniendo el estado y restricción locales intactos (Upsert Híbrido).
      operationId: importMovie
      tags: [Movies]
      security:
        - bearerAuth: []
      parameters:
        - name: tmdbId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '201':
          description: ID local generado (Película guardada)
        '502':
          description: TMDBIntegrationException (Error con la API externa)

  /movies/admin:
    get:
      summary: (Admin) Consultar catálogo local completo
      operationId: getAllMoviesAdmin
      tags: [Movies]
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Catálogo local sin filtros forzados

  /movies/{id}:
    get:
      summary: (Público) Consultar detalles
      operationId: getMovieById
      tags: [Movies]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Documento detallado de la película con metadata JSONB extraída
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MovieDetailResponse'
        '404':
          description: MovieNotFoundException
    put:
      summary: (Admin) Actualizar metadatos y clasificaciones
      operationId: updateMovie
      tags: [Movies]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                overview:
                  type: string
                ageRating:
                  type: string
                status:
                  type: string
      responses:
        '200':
          description: Película actualizada
        '400':
          description: ConstraintViolationException
    delete:
      summary: (Admin) Eliminar película (Soft Delete)
      operationId: deleteMovie
      tags: [Movies]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Eliminación lógica exitosa
        '409':
          description: IllegalMovieStateException (Ya tiene funciones)

  /movies:
    get:
      summary: (Público) Ver cartelera
      operationId: getCartelera
      tags: [Movies]
      parameters:
        - name: genre
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Películas activas en cartelera (Fuerza status=CARTELERA)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MovieCarteleraResponse'

  /movies/pre-estreno:
    get:
      summary: (Público) Ver próximos estrenos en preventa
      operationId: getPreEstrenos
      tags: [Movies]
      parameters:
        - name: genre
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Películas en estado PRE-ESTRENO
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MovieCarteleraResponse'

  # ==========================================
  # SHOWTIME SERVICE
  # ==========================================
  /showtimes:
    get:
      summary: Obtiene funciones disponibles por fecha
      tags: [Showtimes]
      parameters:
        - name: fecha
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Lista de funciones programadas
    post:
      summary: Programa una nueva función
      tags: [Showtimes]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                movie_id:
                  type: integer
                  format: int64
                sala_id:
                  type: string
                  format: uuid
                proyeccion_id:
                  type: integer
                fecha_inicio:
                  type: string
                  format: date-time
                precio_ticket:
                  type: number
      responses:
        '201':
          description: Función programada y asientos pre-generados
        '409':
          description: Solapamiento de horarios en la misma sala

  # ==========================================
  # SEAT SERVICE
  # ==========================================
  /seats:
    get:
      summary: Obtiene el mapa de butacas de una función
      tags: [Seats]
      parameters:
        - name: funcionId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Mapa de butacas y su estado actual

  /seats/lock:
    post:
      summary: Bloquea un asiento temporalmente
      tags: [Seats]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ticket_id:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Asiento bloqueado con éxito
        '409':
          description: Asiento ya fue tomado por otro usuario (OptimisticLockException)